Separated some components based on single responsibility principles
