+++ src/components/ApplicationViews.js
@@ -39,15 +39,9 @@ export const ApplicationViews = () => {
                 <RequestForm />
             </Route>
 
-            <Route exact path="/register">
-                <Register />
-            </Route>
-
             <Route exact path="/login">
                 <Login />
             </Route>
-
-
 
             <Route exact path="/hiring">
                 <Hire />
